# src/main/resources/swagger.yml
...
swagger: "2.0"
info:
  description: Api Documentation
  version: "2.0"
  title: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: localhost:8080
basePath: /
...
paths:
...
  /employees/{id}:
    get:
      tags:
      - employees
      summary: Get an Employee
      operationId: getEmployeeUsingGET
      produces:
      - application/json
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: integer
        format: int32
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Employee'
        404:
          description: Not Found
    delete:
      tags:
      - employees
      summary: Remove an Employee
      operationId: removeEmployeeUsingDELETE
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: integer
        format: int32
      responses:
        204:
          description: No Content
  /add:
    put:
      tags:
      - employees
      summary: Add an Employee
      operationId: addEmployeeUsingPUT
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: employee
        description: employee
        required: true
        schema:
          $ref: '#/definitions/Employee'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/Employee'
...
definitions:
  Employee:
    type: object
    properties:
      createdAt:
        type: string
        example: yyyy-MM-dd@HH:mm:ss.SSSZ
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
        format: int
      lastName:
        type: string
    title: Employee
    example:
      createdAt: yyyy-MM-dd@HH:mm:ss.SSSZ
      firstName: firstName
      lastName: lastName
      id: 0
      email: email